single-agent-system/pyproject.toml
[project]
name = "single-agent-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "databricks-agents>=1.9.0",
    "databricks-langchain>=0.12.1",
    "langchain-google-genai>=4.1.2",
    "langgraph>=1.0.5",
    "langgraph-supervisor>=0.0.31",
    "pydantic-settings>=2.12.0",
]
[tool.uv]
package = true

[dependency-groups]
dev = [
    "databricks-connect>=16.1.7",
    "langgraph-cli[inmem]>=0.4.11",
    "pytest>=8.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
